feat(charts): separate chart for online use and for export
