+++ project/charts/UsagePieChart.py
@@ -1,0 +1,10 @@
+from .CouverturePieChart import CouverturePieChart, CouverturePieChartExport
+
+
+class UsagePieChart(CouverturePieChart):
+    _level = 1
+    _sol = "usage"
+
+
+class UsagePieChartExport(CouverturePieChartExport):
+    pass
