+++ project/charts/__init__.py
@@ -1,0 +1,109 @@
+__all__ = [
+    "AnnualArtifChart",
+    "AnnualArtifChartExport",
+    "AnnualConsoByDeterminantChart",
+    "AnnualConsoByDeterminantChartExport",
+    "AnnualConsoByHouseholdChart",
+    "AnnualConsoByHouseholdChartExport",
+    "AnnualConsoByPopChart",
+    "AnnualConsoByPopChartExport",
+    "AnnualConsoChart",
+    "AnnualConsoChartExport",
+    "AnnualConsoComparisonChart",
+    "AnnualConsoComparisonChartExport",
+    "AnnualConsoProportionalComparisonChart",
+    "AnnualConsoProportionalComparisonChartExport",
+    "AnnualHouseholdChart",
+    "AnnualHouseholdChartExport",
+    "AnnualPopChart",
+    "AnnualPopChartExport",
+    "ArtifByCouverturePieChart",
+    "ArtifByCouverturePieChartExport",
+    "ArtifByUsagePieChart",
+    "ArtifByUsagePieChartExport",
+    "ArtifProgressionByCouvertureChart",
+    "ArtifProgressionByCouvertureChartExport",
+    "ArtifProgressionByUsageChart",
+    "ArtifProgressionByUsageChartExport",
+    "ArtifWaterfallChart",
+    "ArtifWaterfallChartExport",
+    "ConsoByDeterminantPieChart",
+    "ConsoByDeterminantPieChartExport",
+    "CouvertureChangeWheelChart",
+    "CouvertureChangeWheelChartExport",
+    "CouverturePieChart",
+    "CouverturePieChartExport",
+    "CouvertureProgressionChart",
+    "CouvertureProgressionChartExport",
+    "ObjectiveChart",
+    "ObjectiveChartExport",
+    "SurfaceChart",
+    "SurfaceChartExport",
+    "UsageChangeWheelChart",
+    "UsageChangeWheelChartExport",
+    "UsagePieChart",
+    "UsagePieChartExport",
+    "UsageProgressionChart",
+    "UsageProgressionChartExport",
+    "ConsoCommunes",
+    "ConsoCommunesExport",
+    "NetArtifComparaisonChart",
+]
+
+from .AnnualArtifChart import AnnualArtifChart, AnnualArtifChartExport
+from .AnnualConsoByDeterminantChart import (
+    AnnualConsoByDeterminantChart,
+    AnnualConsoByDeterminantChartExport,
+)
+from .AnnualConsoByHouseholdChart import (
+    AnnualConsoByHouseholdChart,
+    AnnualConsoByHouseholdChartExport,
+)
+from .AnnualConsoByPopChart import AnnualConsoByPopChart, AnnualConsoByPopChartExport
+from .AnnualConsoChart import AnnualConsoChart, AnnualConsoChartExport
+from .AnnualConsoComparisonChart import (
+    AnnualConsoComparisonChart,
+    AnnualConsoComparisonChartExport,
+)
+from .AnnualConsoProportionalComparisonChart import (
+    AnnualConsoProportionalComparisonChart,
+    AnnualConsoProportionalComparisonChartExport,
+)
+from .AnnualHouseholdChart import AnnualHouseholdChart, AnnualHouseholdChartExport
+from .AnnualPopChart import AnnualPopChart, AnnualPopChartExport
+from .ArtifByCouverturePieChart import (
+    ArtifByCouverturePieChart,
+    ArtifByCouverturePieChartExport,
+)
+from .ArtifByUsagePieChart import ArtifByUsagePieChart, ArtifByUsagePieChartExport
+from .ArtifProgressionByCouvertureChart import (
+    ArtifProgressionByCouvertureChart,
+    ArtifProgressionByCouvertureChartExport,
+)
+from .ArtifProgressionByUsageChart import (
+    ArtifProgressionByUsageChart,
+    ArtifProgressionByUsageChartExport,
+)
+from .ArtifWaterfallChart import ArtifWaterfallChart, ArtifWaterfallChartExport
+from .ConsoByDeterminantPieChart import (
+    ConsoByDeterminantPieChart,
+    ConsoByDeterminantPieChartExport,
+)
+from .ConsoCommunes import ConsoCommunes, ConsoCommunesExport
+from .CouvertureChangeWheelChart import (
+    CouvertureChangeWheelChart,
+    CouvertureChangeWheelChartExport,
+)
+from .CouverturePieChart import CouverturePieChart, CouverturePieChartExport
+from .CouvertureProgressionChart import (
+    CouvertureProgressionChart,
+    CouvertureProgressionChartExport,
+)
+from .NetArtifComparaisonChart import NetArtifComparaisonChart
+from .ObjectiveChart import ObjectiveChart, ObjectiveChartExport
+from .SurfaceChart import SurfaceChart, SurfaceChartExport
+from .UsageChangeWheelChart import UsageChangeWheelChart, UsageChangeWheelChartExport
+from .UsagePieChart import UsagePieChart, UsagePieChartExport
+from .UsageProgressionChart import UsageProgressionChart, UsageProgressionChartExport
+
+# TODO : remove this line
