+++ project/charts/constants.py
@@ -1,0 +1,21 @@
+DEFAULT_HEADER_FORMAT = "<b>{series.name}</b><br/>"
+DEFAULT_POINT_FORMAT = "{point.name}: {point.y}"
+DEFAULT_VALUE_DECIMALS = 2
+
+CEREMA_CREDITS = {
+    "enabled": True,
+    "text": "Source : Fichiers Fonciers (CEREMA)",
+    "style": {"cursor": "initial"},
+}
+
+OCSGE_CREDITS = {
+    "enabled": True,
+    "text": "Source : OCSGE (IGN)",
+    "style": {"cursor": "initial"},
+}
+
+INSEE_CREDITS = {
+    "enabled": True,
+    "text": "Source : INSEE",
+    "style": {"cursor": "initial"},
+}
