+++ project/templates/project/report_consommation.html
@@ -374,7 +374,6 @@ Rapport consommation
 
         <div class="fr-mt-7w">
             <h3 id="territoires-de-comparaison">Comparaison avec les territoires similaires</h3>
-            <p>Proposition de territoires de même maille administrative, il est possible de modifier cette selection dans la légende</p>
             <div class="border fr-p-2w">
                 <div class="d-flex justify-content-end align-items-center fr-mb-2w">
                     <a class="fr-mr-2w fr-btn fr-icon-add-circle-line fr-btn--tertiary fr-btn--sm" href="{% url 'project:lookalike' project.id %}?next={% url 'project:report_conso' project.id %}&anchor=territoires-de-comparaison" data-toggle="tooltip" title="Ajouter un territoire de comparaison" aria-hidden="true"></a>
